ML/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
deepface==0.0.79
opencv-python-headless==4.7.0.72
numpy==1.26.4
cloudinary==1.44.1
python-dotenv==1.0.0
pillow==10.1.0
